data-services-layer/users/requirements.txt
Flask==3.0.0
gunicorn==20.1.0
pymongo==4.5.0
PyJWT==2.8.0